+++ developer/src/kmc-analyze/package.json
@@ -16,14 +16,15 @@
   "bugs": {
     "url": "https://github.com/keymanapp/keyman/issues"
   },
-  "main": "build/src/index.js",
+  "exports": {
+    ".": "./build/src/index.js"
+  },
   "files": [
     "build/src/"
   ],
   "dependencies": {
     "@keymanapp/common-types": "*",
-    "@keymanapp/kmc-kmn": "*",
-    "commander": "^10.0.0"
+    "@keymanapp/kmc-kmn": "*"
   },
   "devDependencies": {
     "@keymanapp/resources-gosh": "*",
@@ -33,7 +34,6 @@
     "c8": "^7.12.0",
     "chai": "^4.3.4",
     "chalk": "^2.4.2",
-    "esbuild": "^0.15.8",
     "mocha": "^8.4.0",
     "ts-node": "^9.1.1",
     "typescript": "^4.9.5"
