+++ developer/src/kmc-analyze/src/index.ts
@@ -4,3 +4,4 @@
 
 export { AnalyzeOskCharacterUse } from './osk-character-use/index.js';
 export { AnalyzeOskRewritePua } from './osk-rewrite-pua/index.js';
+export { AnalyzerMessages } from './messages.js';
