+++ developer/src/kmc-ldml/src/main.ts
@@ -4,7 +4,7 @@ export { LdmlKeyboardKeymanWebCompiler } from './compiler/keymanweb-compiler.js'
 export { LdmlKeyboardVisualKeyboardCompiler } from './compiler/visual-keyboard-compiler.js';
 export { TouchLayoutCompiler } from './compiler/touch-layout-compiler.js';
 export { LdmlCompilerOptions } from './compiler/ldml-compiler-options.js';
-export { CompilerMessages } from './compiler/messages.js';
+export { CompilerMessages as LdmlKeyboardCompilerMessages } from './compiler/messages.js'; // TODO: rename messages.js and CompilerMessages
 export { KMXPlusMetadataCompiler } from './compiler/metadata-compiler.js';
 
 export { KMXBuilder } from "@keymanapp/common-types";
