+++ developer/src/kmc-model/src/main.ts
@@ -1,1 +1,2 @@
 export { LexicalModelCompiler } from './lexical-model-compiler.js';
+export { ModelCompilerMessages } from './model-compiler-messages.js';
