+++ developer/src/kmc-package/src/compiler/package-version-validator.ts
@@ -1,5 +1,5 @@
 import { KmpJsonFile, CompilerCallbacks, KpsFile } from '@keymanapp/common-types';
-import { CompilerMessages } from './messages.js';
+import { CompilerMessages } from './package-compiler-messages.js';
 import { KeyboardMetadataCollection } from './package-metadata-collector.js';
 
 export const DEFAULT_KEYBOARD_VERSION = '1.0';
