+++ developer/src/kmc-package/src/main.ts
@@ -1,3 +1,4 @@
 export { KmpCompiler } from "./compiler/kmp-compiler.js";
 export { PackageValidation } from "./compiler/package-validation.js";
 export { WindowsPackageInstallerSources, WindowsPackageInstallerCompiler } from "./compiler/windows-package-installer-compiler.js";
+export { CompilerMessages as PackageCompilerMessages } from './compiler/package-compiler-messages.js';
